requirements.txt
-e ./common
-r ./common/requirements-ml.txt 

streamlit
pandas
numpy
plotly
google-cloud-resource-manager 